#164 Move CollisionMap to level package
This removes the dependency from util to level (Track/AbstractEntity)

<organization,logic
