+++ src/main/java/nl/tudelft/ti2206/group9/level/CollisionHandler.java
@@ -5,9 +5,7 @@ import nl.tudelft.ti2206.group9.level.entity.AbstractObstacle;
 import nl.tudelft.ti2206.group9.level.entity.AbstractPickup;
 import nl.tudelft.ti2206.group9.level.entity.Player;
 import nl.tudelft.ti2206.group9.level.entity.PowerupInvulnerable;
-import nl.tudelft.ti2206.group9.util.CollisionMap;
 import nl.tudelft.ti2206.group9.util.GameObserver;
-
 import static nl.tudelft.ti2206.group9.ShaftEscape.OBSERVABLE;
 
 /**
